wsd: sign-to-unsigned conversion
net/Socket.hpp:1315:22: runtime error: implicit conversion from type
'ssize_t' (aka 'long') of value -1 (64-bit, signed) to type 'unsigned
long' changed the value to 18446744073709551615 (64
-bit, unsigned)
SUMMARY: UndefinedBehaviorSanitizer: undefined-behavior
net/Socket.hpp:1315:22

+++ net/Socket.hpp
@@ -1312,7 +1312,7 @@ public:
                     _inBuffer.append(&buf[0], len);
                 }
                 // else poll will handle errors.
-            } while (len == (sizeof(buf)));
+            } while (len == static_cast<ssize_t>(sizeof(buf)));
 
             // Restore errno from the read call.
             errno = last_errno;
